show proposals on top on main screen
