+++ Example/WalletApp/PresentationLayer/Wallet/Main/MainModule.swift
@@ -5,7 +5,8 @@ final class MainModule {
     @discardableResult
     static func create(app: Application) -> UIViewController {
         let router = MainRouter(app: app)
-        let presenter = MainPresenter(router: router)
+        let interactor = MainInteractor()
+        let presenter = MainPresenter(router: router, interactor: interactor)
         let viewController = MainViewController(presenter: presenter)
 
         router.viewController = viewController
