+++ Example/WalletApp/PresentationLayer/Wallet/Main/MainPresenter.swift
@@ -3,7 +3,9 @@ import Combine
 
 final class MainPresenter {
 
+    private let interactor: MainInteractor
     private let router: MainRouter
+    private var disposeBag = Set<AnyCancellable>()
 
     var tabs: [TabPage] {
         return TabPage.allCases
@@ -16,7 +18,31 @@ final class MainPresenter {
         ]
     }
 
-    init(router: MainRouter) {
+    init(router: MainRouter,
+         interactor: MainInteractor) {
+        defer {
+            setupInitialState()
+        }
         self.router = router
+        self.interactor = interactor
     }
 }
+
+// MARK: - Private functions
+extension MainPresenter {
+    private func setupInitialState() {
+        interactor.pushRequestPublisher
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] request in
+                self?.router.present(pushRequest: request)
+            }.store(in: &disposeBag)
+
+        interactor.sessionProposalPublisher
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] proposal in
+                self?.router.present(proposal: proposal)
+            }
+            .store(in: &disposeBag)
+    }
+}
+
