+++ Example/WalletApp/PresentationLayer/Wallet/Wallet/WalletInteractor.swift
@@ -7,17 +7,9 @@ final class WalletInteractor {
     var requestPublisher: AnyPublisher<AuthRequest, Never> {
         return Web3Wallet.instance.authRequestPublisher
     }
-    
-    var sessionProposalPublisher: AnyPublisher<Session.Proposal, Never> {
-        return Web3Wallet.instance.sessionProposalPublisher
-    }
-    
+
     var sessionRequestPublisher: AnyPublisher<Request, Never> {
         return Web3Wallet.instance.sessionRequestPublisher
-    }
-    
-    var pushRequestPublisher: AnyPublisher<(id: RPCID, account: Account, metadata: AppMetadata), Never> {
-        return Push.wallet.requestPublisher
     }
 
     var sessionsPublisher: AnyPublisher<[Session], Never> {
