+++ Example/WalletApp/PresentationLayer/Wallet/Wallet/WalletRouter.swift
@@ -1,7 +1,5 @@
 import UIKit
-
 import Web3Wallet
-import WalletConnectPush
 
 final class WalletRouter {
     weak var viewController: UIViewController!
@@ -16,19 +14,9 @@ final class WalletRouter {
         AuthRequestModule.create(app: app, request: request)
             .presentFullScreen(from: viewController, transparentBackground: true)
     }
-    
-    func present(proposal: Session.Proposal) {
-        SessionProposalModule.create(app: app, proposal: proposal)
-            .presentFullScreen(from: viewController, transparentBackground: true)
-    }
-    
+
     func present(sessionRequest: Request) {
         SessionRequestModule.create(app: app, sessionRequest: sessionRequest)
-            .presentFullScreen(from: viewController, transparentBackground: true)
-    }
-
-    func present(pushRequest: PushRequest) {
-        PushRequestModule.create(app: app, pushRequest: pushRequest)
             .presentFullScreen(from: viewController, transparentBackground: true)
     }
